adding lodestar-vesting
